Refactor : 폴더 구조 변경 - 회의내용에 맞춰 폴더구조를 각 기능단위와 shared로 구분함
